+++ example/lib/page/home.dart
@@ -1,10 +1,11 @@
 import 'package:provider/provider.dart';
-import 'package:fhel_example/page/utils.dart';
 import 'package:flutter/material.dart';
 import 'package:fhel_example/addition.dart';
 import 'user_input.dart';
 import 'bottom_bar.dart';
 import 'settings.dart';
+import 'utils.dart';
+import 'log.dart';
 
 class HomePage extends StatelessWidget {
   final _tab1navigatorKey = GlobalKey<NavigatorState>();
@@ -27,56 +28,6 @@ class HomePage extends StatelessWidget {
           navigatorkey: _tab3navigatorKey,
         ),
       ],
-    );
-  }
-}
-
-class LogView extends ChangeNotifier {
-  List<String> logs = [];
-
-  void log(String log) {
-    logs.add(log);
-    notifyListeners();
-  }
-}
-
-class LogScreen extends StatefulWidget {
-  // final Function(String) addLogCallback;
-
-  const LogScreen({super.key});
-
-  @override
-  LogListScreen createState() {
-    return LogListScreen();
-  }
-}
-
-bool isException(String log) {
-  return log.contains('Exception');
-}
-
-class LogListScreen extends State<LogScreen> {
-  @override
-  Widget build(BuildContext context) {
-    var logs = Provider.of<SessionChanges>(context).logs;
-    // in reverse order
-    logs = logs.reversed.toList();
-    return SizedBox(
-      height: 365,
-      child: ListView.builder(
-        itemCount: logs.length,
-        itemBuilder: (context, index) {
-          return ListTile(
-            title: Text(
-              logs[index],
-              style: TextStyle(
-                // fontSize: 14,
-                color: isException(logs[index]) ? Colors.red : Colors.black,
-              ),
-            ),
-          );
-        },
-      ),
     );
   }
 }
